// nextjs/src/types/booking/bookings-status.enum.ts
export enum BookingStatusEnum {
    COMPLETED = 'completed',
    PENDING = 'pending',
    SCHEDULED = 'scheduled',
    CANCELED_BY_TRAINER = 'canceled_by_trainer',
    CANCELED_BY_CLIENT = 'canceled_by_client',
    CANCELED_BY_SYSTEM = 'canceled_by_system',
    CANCELED_PAYMENT_FAILED = 'canceled_payment_failed',
}